TEST: test de la fonction getLines
