+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "vue-tsc -b && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "vitest"
   },
   "dependencies": {
     "vue": "^3.5.13"
@@ -20,5 +21,5 @@
     "vite": "^6.0.5",
     "vitest": "^3.0.7",
     "vue-tsc": "^2.2.0"
-  }
+  }  
 }
